% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibBarplot.R
\name{hsBarplot2}
\alias{hsBarplot2}
\title{barplot() with X-Position Support (2)}
\usage{
hsBarplot2(
  myHeight,
  myPos = NULL,
  myWidth = 1,
  myBeside = FALSE,
  myInnerSpace = 0,
  myValLabs = TRUE,
  dx = 0,
  dy = 0,
  dbg = FALSE,
  ...
)
}
\arguments{
\item{myHeight}{Vector or matrix containing the bar heights, see R's barplot
function.}

\item{myPos}{Vector containing the x-positions of the bars/bar groups.}

\item{myWidth}{Bar widths. If bars are to be stacked (isp = NULL, see below)
w represents the widths of the stacked bars. If bars shall be plotted side
by side instead of stacked (isp >= 0) w represents the widths of the bars
within one and the same bar group. Corresponding bars of different groups
will always have the same width.}

\item{myBeside}{If TRUE, bars within a bar group are arranged side by side
instead of stacked.}

\item{myInnerSpace}{"Inner" space between bars of the same bar group (only
relevant if y is a matrix and not a vector). If NULL, bars of the same bar
group will be stacked}

\item{myValLabs}{whether to put value labels on top of the bars}

\item{dx}{x offset for text labels. Default: 0}

\item{dy}{y offset for text labels. Default: 0}

\item{dbg}{whether to print debug messages or not}

\item{\dots}{further arguments passed to \code{\link{barplot}}}
}
\description{
Extended version of R barplot function, enabling x positioning of the bars.
}
