% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibBarplot.R
\name{hsBpBargroupWidth}
\alias{hsBpBargroupWidth}
\title{Width of Bargroup}
\usage{
hsBpBargroupWidth(myWidth, myBarsPerGroup, myInnerSpace)
}
\arguments{
\item{myWidth}{widths of the bars in one group}

\item{myBarsPerGroup}{number of bars per group}

\item{myInnerSpace}{inner space between bars in a bar group}
}
\value{
vector of numeric of length one
}
\description{
The widths represent the widths of the bars in one group. It is calculated
as: sum(myWidth) + (myBarsPerGroup - 1) * myInnerSpace
}
