% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibBarplot.R
\name{hsBpDefaultXlim}
\alias{hsBpDefaultXlim}
\title{Default xlim for Barplot}
\usage{
hsBpDefaultXlim(position, barGroupWidth = 1)
}
\arguments{
\item{position}{vector of x-positions of bars}

\item{barGroupWidth}{bargroup widths. Default: 1}
}
\value{
vector of numeric of length two
}
\description{
Default xlim for Barplot
}
