% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibBarplot.R
\name{hsBpGetXlim}
\alias{hsBpGetXlim}
\title{xlim for Barplot}
\usage{
hsBpGetXlim(myData, myWidth = 1, mySpace = NULL, myBeside = FALSE)
}
\arguments{
\item{myData}{corresponding to "height" in \code{\link{barplot}}}

\item{myWidth}{corresponding to "width" in \code{\link{barplot}}}

\item{mySpace}{corresponding to "space" in \code{\link{barplot}}}

\item{myBeside}{corresponding to "beside" in \code{\link{barplot}}}
}
\value{
vector of numeric of length two
}
\description{
Depending on the argument values that are passed to \code{barplot()} a good
vector to be used as xlim is returned.
}
