% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibBarplot.R
\name{hsBpNumBars}
\alias{hsBpNumBars}
\title{Number of Bar Groups and Bars Per Group Produced by barplot()}
\usage{
hsBpNumBars(myData, myBeside)
}
\arguments{
\item{myData}{as passed in \code{height} to \code{\link{barplot}}}

\item{myBeside}{as passed in \code{beside} to \code{\link{barplot}}}
}
\value{
Returns tupel containing N = number of bar groups and n = number of
bars per group
}
\description{
Number of bar groups (N) and bars per group (n) produced by barplot(),
depending on given data and parameter "beside"
}
