% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibBarplot.R
\name{hsPutValueLabels}
\alias{hsPutValueLabels}
\title{Put Values on Top of the Bars}
\usage{
hsPutValueLabels(myHeight, myPosition, myWidth, mySpace, myYmax)
}
\arguments{
\item{myHeight}{heights of bars}

\item{myPosition}{x-positions of bars}

\item{myWidth}{widths of bars}

\item{mySpace}{space between bars in a bar group}

\item{myYmax}{y limit. Used to determine the distance between top of bar and
labels on top}
}
\description{
Put Values on Top of the Bars
}
